started P3 material file parsing.

+++ src/main/cs410/raytracer/RGB.java
@@ -5,17 +5,17 @@ public class RGB {
     public final static int r = 0;
     public final static int g = 1;
     public final static int b = 2;
-    int[] rgb;
+    float[] rgb;
     
     /**
      * Creates a new black pixel
      */
     public RGB(){
-        rgb = new int[3];
+        rgb = new float[3];
     }
     
-    public RGB(int r, int g, int b){
-        rgb = new int[3];
+    public RGB(float r, float g, float b){
+        rgb = new float[3];
         rgb[RGB.r] = r;
         rgb[RGB.g] = g;
         rgb[RGB.b] = b;
